Make use of pkg_resources module for static files
Instead of relying on relative paths, pkg_resources can be used to
retrieve static files/data in a more reliable manner. This commit
therefore moves the current .json files for the quizzes into a static
folder which is automatically packaged by setup.py. These files can then
be used by running the `pkg_resources.resource_filename` command. The
main advantage of this is that it will work regardless of where the
EduBot package is installed.

Summary of Changes:
- Move cogs/poll.py to src/edubot/cogs/poll.py
- Move .json files from cogs/quizzes to src/edubot/static/quizzes
- Replace relative path with call to pkg_resources.resource_filename