Create queue from channel history
The #assignment-hand-in channels of A and B have a large queue
already filled. Simply discarding those to make use of the bot
queues would work, but this commit should be able to make one
from the existing messages.
Filters first on readyness, then on whether people have already
been checked. It won't delete non-student messages, but will
clean up the rest as much as possible.